+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * API: allow setting `text_range` when creating comments via API
 * Fix duplicate entries in template search results
 * Allow searching for member usernames in project list
+* Enable toggle comment shortcut in HTML/CSS editor
 
 
 ## v2025.20 - 2025-03-05